station service updated

+++ mcsv-station/src/main/java/com/arqui/integrador/repository/IStationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.arqui.integrador.model.Station;
 
-public interface IStationRepository extends JpaRepository<Station, Integer>{
+public interface IStationRepository extends JpaRepository<Station, Long>{
 	
 }
